third_party/blink/web_tests/flag-specific/highdpi/external/wpt/css/css-sizing/contain-intrinsic-size/contain-intrinsic-size-029-expected.txt

This is a testharness.js-based test.
[FAIL] .test 5
  assert_equals: \n<button class="test cis-none min-size" data-expected-client-width="20" data-expected-client-height="5"></button>\nclientWidth expected 20 but got 21
[FAIL] .test 6
  assert_equals: \n<button class="test cis-height min-size" data-expected-client-width="20" data-expected-client-height="50"></button>\nclientWidth expected 20 but got 21
[FAIL] .test 7
  assert_equals: \n<button class="test cis-width min-size" data-expected-client-width="100" data-expected-client-height="5"></button>\nclientHeight expected 5 but got 6
Harness: the test ran to completion.
